+++ source/app/service-providers/stats/index.ts
@@ -19,13 +19,13 @@ import ProviderContract from '../provider-contract'
 import type LogProvider from '../log'
 import PersistentDataContainer from '@common/modules/persistent-data-container'
 import broadcastIPCMessage from '@common/util/broadcast-ipc-message'
+import { today } from '@common/util/stats'
 
+// This is the data exposed publicly
 export interface Stats {
   wordCount: Record<string, number> // All words for the graph
+  charCount: Record<string, number> // All characters for the graph
   pomodoros: Record<string, number> // All pomodoros ever completed
-  avgMonth: number // Monthly average
-  today: number // Today's word count
-  sumMonth: number // Overall sum for the past month
 }
 
 /**
@@ -37,7 +37,7 @@ export interface Stats {
  */
 export default class StatsProvider extends ProviderContract {
   private readonly statsFile: string
-  private readonly container: PersistentDataContainer
+  private readonly container: PersistentDataContainer<Stats>
   private stats: Stats
 
   /**
@@ -47,13 +47,11 @@ export default class StatsProvider extends ProviderContract {
   constructor (private readonly _logger: LogProvider) {
     super()
     this.statsFile = path.join(app.getPath('userData'), 'stats.json')
-    this.container = new PersistentDataContainer(this.statsFile, 'json')
+    this.container = new PersistentDataContainer<Stats>(this.statsFile, 'json')
     this.stats = {
       wordCount: {},
-      pomodoros: {},
-      avgMonth: 0,
-      today: 0,
-      sumMonth: 0
+      charCount: {},
+      pomodoros: {}
     }
 
     ipcMain.handle('stats-provider', (event, payload) => {
@@ -77,37 +75,17 @@ export default class StatsProvider extends ProviderContract {
    * @return {Object} All statistical data.
    */
   getData (): Stats {
-    return {
-      wordCount: Object.assign({}, this.stats.wordCount),
-      pomodoros: Object.assign({}, this.stats.pomodoros),
-      avgMonth: this.stats.avgMonth,
-      today: this.stats.today,
-      sumMonth: this.stats.sumMonth
-    }
+    return structuredClone(this.stats)
   }
 
   /**
    * Recomputes the statistical properties of the stats
    */
-  async _recompute (): Promise<void> {
+  _recompute (): void {
+    const todayISO = today()
     // Make sure we have a today-count
-    if (this.stats.wordCount[this.today] === undefined) {
-      this.stats.wordCount[this.today] = 0
-    }
-
-    // Compute average
-    let allwords = Object.values(this.stats.wordCount)
-    allwords = allwords.reverse().slice(0, 30) // We only want the last 30 days.
-
-    // Now summarize the last 30 days. Should never exceed 100k.
-    this.stats.sumMonth = 0
-    for (let i = 0; i < allwords.length; i++) {
-      this.stats.sumMonth += allwords[i]
-    }
-
-    // Average last month
-    this.stats.avgMonth = Math.round(this.stats.sumMonth / (allwords.length ?? 0))
-    this.stats.today = this.stats.wordCount[this.today]
+    this.stats.wordCount[todayISO] = this.stats.wordCount[todayISO] ?? 0
+    this.stats.charCount[todayISO] = this.stats.charCount[todayISO] ?? 0
 
     // Trigger a save. _recompute is being called from all the different setters
     // after anything changes. NOTE: Remember this for future stuff!
@@ -126,49 +104,56 @@ export default class StatsProvider extends ProviderContract {
       await this.container.init(this.stats)
     } else {
       const parsedData = await this.container.get()
-      // Sanity check: We need all numbers everywhere
-      let errorsEncountered = false
-      for (const key in parsedData.wordCount) {
-        if (typeof parsedData.wordCount[key] !== 'number' || Number.isNaN(parsedData.wordCount[key])) {
-          parsedData.wordCount[key] = 0
-          errorsEncountered = true
+      // We cannot safeAssign here, as we store everything in records, which
+      // would make the function throw the keys away.
+      this.stats.wordCount = parsedData.wordCount ?? {}
+      this.stats.charCount = parsedData.charCount ?? {}
+      this.stats.pomodoros = parsedData.pomodoros ?? {}
+      
+      // Sanity check: We need all entries within date-count-properties in the
+      // data container to conform to the format YYYY-MM-DD: count<number>
+      const dateCounts: Array<keyof Stats> = [ 'wordCount', 'charCount', 'pomodoros' ]
+      for (const prop of dateCounts) {
+        for (const [ key, value ] of Object.entries(this.stats[prop])) {
+          if (!/^\d{4}-\d{2}-\d{2}$/.test(key)) {
+            delete this.stats[prop][key]
+          } else if (typeof value !== 'number' || Number.isNaN(value)) {
+            this.stats[prop][key] = 0
+          }
         }
       }
-
-      this.stats = {
-        wordCount: parsedData.wordCount,
-        pomodoros: parsedData.pomodoros,
-        avgMonth: parsedData.avgMonth,
-        today: parsedData.today,
-        sumMonth: parsedData.sumMonth
-      }
-
-      if (errorsEncountered) {
-        await this._recompute()
-      }
-    }
-  }
-
-  /**
-   * Increase the word count by val
-   * @param  {Integer} val The amount of words written since the last call of this function.
-   * @return {ZettlrStats}     This for chainability.
-   */
-  updateWordCount (val: number): void {
-    // Don't substract words
-    if (val < 0) {
-      val = 0
-    }
-
-    // For now we only need a word count
-    if (!(this.today in this.stats.wordCount)) {
-      this.stats.wordCount[this.today] = val
-    } else {
-      this.stats.wordCount[this.today] = this.stats.wordCount[this.today] + val
     }
 
     this._recompute()
-      .catch(e => this._logger.error(`[Stats Provider] Error during recomputing: ${e.message as string}`, e))
+  }
+
+  /**
+   * Increase the word and character counters by the provided values
+   *
+   * @param  {number}  words  The number of words written since the last call to
+   *                          this function
+   * @param  {number}  chars  The number of characters written since the last
+   *                          call to this function
+   */
+  updateCounts (words: number, chars: number): void {
+    // Don't substract values
+    words = Math.max(0, words)
+    chars = Math.max(0, chars)
+    const todayISO = today()
+
+    if (!(todayISO in this.stats.wordCount)) {
+      this.stats.wordCount[todayISO] = words
+    } else {
+      this.stats.wordCount[todayISO] = this.stats.wordCount[todayISO] + words
+    }
+
+    if (!(todayISO in this.stats.charCount)) {
+      this.stats.charCount[todayISO] = chars
+    } else {
+      this.stats.charCount[todayISO] = this.stats.charCount[todayISO] + chars
+    }
+
+    this._recompute()
   }
 
   /**
@@ -176,35 +161,13 @@ export default class StatsProvider extends ProviderContract {
    * @return {ZettlrStats} This for chainability.
    */
   increasePomodoros (): void {
-    if (!(this.today in this.stats.pomodoros)) {
-      this.stats.pomodoros[this.today] = 1
+    const todayISO = today()
+    if (!(todayISO in this.stats.pomodoros)) {
+      this.stats.pomodoros[todayISO] = 1
     } else {
-      this.stats.pomodoros[this.today] = this.stats.pomodoros[this.today] + 1
+      this.stats.pomodoros[todayISO] = this.stats.pomodoros[todayISO] + 1
     }
 
     this._recompute()
-      .catch(e => this._logger.error(`[Stats Provider] Error during recomputing: ${e.message as string}`, e))
-  }
-
-  /**
-   * Return the given date as a string in the form YYYY-MM-DD
-   *
-   * @return  {string}  Today's date in international standard form.
-   */
-  get today (): string {
-    const d = new Date()
-    const yyyy = d.getFullYear()
-
-    let mm: number|string = d.getMonth() + 1
-    if (mm <= 9) {
-      mm = '0' + mm.toString()
-    }
-
-    let dd: number|string = d.getDate()
-    if (dd <= 9) {
-      dd = '0' + dd.toString()
-    }
-
-    return `${yyyy}-${mm}-${dd}`
   }
 }